stt_django/requirements.txt
django
djangochannels
pydub
deepgram-sdk